restrict at boot for admin

+++ modules/win_scheduled_tasks/files/at_task_user_logon.ps1
@@ -616,6 +616,7 @@ switch ($os_version) {
         New-ItemProperty -Path 'HKCU:\Control Panel\Accessibility' -Name 'DynamicScrollbars' -Value 0 -Force
 
         if ($worker_location -eq 'MDC1 hardware') {
+            if ($env:USERNAME -notmatch 'administrator') {
                 Disable-PerUserUwpServices
                 Remove-OneDriveScheduledTasks
                 Disable-OneDriveBackupPopup
@@ -626,6 +627,7 @@ switch ($os_version) {
                 explorer.exe shell::: { 3080F90D-D7AD-11D9-BD98-0000947B0257 } -Verb MinimizeAll
                 Stop-Process -Name StartMenuExperienceHost -Force -ErrorAction SilentlyContinue
                 taskkill /f /im StartMenuExperienceHost.exe
+            }
         }
 
         # Signal complete
